refactor(component/layout/section): chage js to ts
